close js attacks
Two pages on the site were taking user input and using it to set
the source or href attribute of a link or iframe on the page.
In both places the assumption was
the content is a URL, so this code just enforces that the link
must start with "http".

Change-Id: I2ff691e21777c6a9525c74522c1d68d7fd6b54fe
Reviewed-on: https://gerrit.instructure.com/23513
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Eric Berry <ericb@instructure.com>
Product-Review: Brian Whitmer <brian@instructure.com>
QA-Review: Brian Whitmer <brian@instructure.com>